requirements.txt
torch~=1.11.0
transformers~=4.18.0
pytorch-crf~=0.7.2
pandas~=1.3.5
scikit-learn~=1.0.2
vncorenlp